[framework] implement immovable objects
Make it so that objects can be made immovable via the constructor ref.
Even if other entities have access to the constructor ref.

Add the functionality to fungible asset to make immovable stores by
default for assets that opt into the behavior.

Also offer a root_owner function to determine the deepest node that owns
a nested object.